+++ rog-aura/src/aura_detection.rs
@@ -1,11 +1,20 @@
 use log::{error, info, warn};
 use serde_derive::{Deserialize, Serialize};
 
+use crate::usb::AuraDevice;
 use crate::{AdvancedAuraType, AuraModeNum, AuraZone};
 
 pub const ASUS_LED_MODE_CONF: &str = "/usr/share/asusd/aura_support.ron";
 pub const ASUS_LED_MODE_USER_CONF: &str = "/etc/asusd/asusd_user_ledmodes.ron";
-pub const ASUS_KEYBOARD_DEVICES: [&str; 6] = ["1866", "18c6", "1869", "1854", "19b6", "1a30"];
+pub const ASUS_KEYBOARD_DEVICES: [AuraDevice; 7] = [
+    AuraDevice::Tuf,
+    AuraDevice::X1854,
+    AuraDevice::X1869,
+    AuraDevice::X1866,
+    AuraDevice::X18c6,
+    AuraDevice::X19b6,
+    AuraDevice::X1a30,
+];
 
 #[derive(Debug, Default, Clone, PartialEq, Eq, Deserialize, Serialize)]
 pub struct LedSupportFile(Vec<LaptopLedData>);